requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
click==8.1.3
Django==3.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
importlib-resources==5.10.1
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
psycopg2-binary==2.9.1
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
zipp==3.11.0
